requirements.txt
psycopg2-binary
pytest
pytest-logging
python-dotenv